requirements.txt
streamlit
python-dotenv
groq
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf